Refactored test code
