--- tests2/Misct/test_webconnections.py
@@ -1,88 +1,0 @@
-import pytest
-import configparser
-import os
-from misc.webConnections import WebConnect
-from PyQt5.QtCore import QSettings
-
-@pytest.fixture(scope="function")
-def web_connect(tmpdir):
-    """Fixture to provide an instance of WebConnect with a temporary QGIS3.ini file."""
-    qini_fp = tmpdir.join("QGIS3.ini")
-    qini_fp.write("[connections]\n")  
-
-    web_connect = WebConnect(qini_fp=str(qini_fp))
-
-    settings = QSettings(str(qini_fp), QSettings.IniFormat)
-    settings.clear() 
-    settings.sync()
-
-    yield web_connect
-
-@pytest.fixture(scope="function")
-def config_file(tmpdir, request):
-    """Fixture to create a temporary config file with the specified content."""
-    test_cases = {
-        "valid_wms_connection": {
-            "config_content": {
-                "AutomaticallyExtractedBuildings": {
-                    "group": "connections-wms\\NRCan_AutomaticallyExtractedBuildings",
-                    "url": "https://maps.geogratis.gc.ca/wms/automatic_extraction_building_footprint_en?request=getcapabilities&service=wms&layers=automatic_extraction_building_footprint_en&version=1.3.0&legend_format=image/png&feature_info_type=text/html"
-                }
-            },
-            "expected_key": "connections/ows/items/wms/connections/items/AutomaticallyExtractedBuildings/url"
-        },
-        "valid_wcs_connection": {
-            "config_content": {
-                "GAR15": {
-                    "group": "connections-wcs\\UNISDR_GAR15_GlobalRiskAssessment",
-                    "url": "http://preview.grid.unep.ch/geoserver/wcs?bbox=-180,-89,180,84&styles=&version=1.0.0&coverage=GAR2015:flood_hazard_1000_yrp&width=640&height=309&crs=EPSG:4326"
-                }
-            },
-            "expected_key": "connections/ows/items/wcs/connections/items/GAR15/url"
-        },
-        "valid_arcgis_connection": {
-            "config_content": {
-                "NPRI": {
-                    "group": "connections-arcgisfeatureserver\\ECCC_NationalPollutantReleaseInventory_NPRI",
-                    "url": "https://maps-cartes.ec.gc.ca/arcgis/rest/services/STB_DGST/NPRI/MapServer"
-                }
-            },
-            "expected_key": "connections/arcgisfeatureserver/items/NPRI/url"
-        }
-    }
-
-    config = configparser.ConfigParser()
-    config.read_dict(test_cases[request.param]["config_content"])
-    config_file_path = tmpdir.join("test_config.ini")
-    with open(config_file_path, 'w') as configfile:
-        config.write(configfile)
-    
-    return str(config_file_path), test_cases[request.param]["expected_key"]
-
-@pytest.mark.parametrize(
-    "config_file, expected_key",
-    [
-        ("valid_wms_connection", "connections/ows/items/wms/connections/items/AutomaticallyExtractedBuildings/url"),
-        ("valid_wcs_connection", "connections/ows/items/wcs/connections/items/GAR15/url"),
-        ("valid_arcgis_connection", "connections/arcgisfeatureserver/items/NPRI/url") 
-    ],
-    indirect=["config_file"]
-)
-def test_read_connections(web_connect, config_file, expected_key):
-    """Test the read_connections function and check expected QGIS3.ini structure."""
-    config_path, expected_key = config_file
-
-    # Run the connection update
-    web_connect.read_connections(config_path)
-
-    # Force QSettings to write changes immediately
-    settings = QSettings(web_connect.qini_fp, QSettings.IniFormat)
-    settings.sync()  # Ensure settings are written to disk
-
-    qini_fp = web_connect.qini_fp
-    with open(qini_fp, "r") as f:
-        ini_content = f.read()
-
-    all_keys = settings.allKeys()
-
-    assert expected_key in all_keys, f"Expected key '{expected_key}' was not found in the updated QGIS3.ini file"
